ML-APi/Video-summarizer/requirements.txt
youtube_transcript_api
google-generativeai
python-dotenv
pathlib
fastapi
pydantic
google-cloud-translate
